fix: parallel bitext tasks scores
Due to the implementation of parallel bitext tasks it computes all combinations even though only some are specified by eval_langs. This leads to a minor computational overhead when evaluating and most notably an error when constructing MTEBResults.

This fix does the following:
- it ensures that only desired language pairs are being evaluated
- it simplified bitext evaluation (there is no longer two cases)
- it additionally only encodes the sentence pairs required for the specific selected pairs (make it faster when we only want to consider a subset of languages)

I tested it using:
- A monolingual bitext task
- A multilingual parallel bitext task
- A multilingal non-parallel bitext task

All pass and reproduce the same results (cheers!)