mock implementation of sparse Encoders
Implements a naïve support for some for sparse encoders. The goal here is to capture some of the problems we have to solve.

Generally it seems like the we can't slice or index a sparse tensor and thus have to implement wrappers for these.

When we hand the vector off to a classfier we have to convert it to dense (even though the classifier itself does support sparse matrices, e.g. for logistic regression).