fix: Seperate model implementations and encoder specifications
- Move all implementations into a seperate folder called `model_implementations`
- moved `encoder_interface.py` and `model_meta.py` into `models`
- renamed `models/*` to `encoder_implementions/*` to make the distinction from between the two folder clear
- merged `models/utils.py` into the only model that used it

We seems to have a few differing names when referring to a model (ModelMeta, get_model, etc.) and encoders (Encoder, AbsEncoder). Should we try to do something about this or just leave it as is?

There is also an inconsistency between how tasks and implementations are in seperate folders, but for benchmark this is not the case.

We could convert it to:
```
benchmarks/tasks/models
| - implementations/*
| - ... # definitions utilities etc.
```
But I am not sure it is worth it and for tasks it might be too much nesting. So I would probably leave it as is.

Note: There is a few refactors that I would like to do on top of this, but will add that to seperate PR (since it is too hard to review here)

Fixes #2299