[v2] don't import transformers on intialization
set_seed is called during the initialization of the aggregate tasks. I have refactored it to take the main components from transformers set_seed. We do not initialize the seed for specialized hardware, but maybe that is also not needed?

With this change we no longer need make transformers an optional install.

This give a notably increase in load time
6.511, 6.798, 6.423
to
5.213, 5.230, 4.929